Move interfaces closer to implementation of extension methods, add missing comment to replace-nudge, do not strip comments in distributed js
