+++ codemods/src/utils/__tests__/register-collection-extensions.test.ts
@@ -1,4 +1,4 @@
-import jscodeshift, { ASTPath, JSXElement, JSXIdentifier } from 'jscodeshift'
+import jscodeshift, { ASTPath, JSXIdentifier } from 'jscodeshift'
 import { registerCollectionExtensions } from '../register-collection-extensions'
 import { stripIndent } from 'common-tags'
