+++ codemods/src/utils/register-import-declaration-collection-extensions.ts
@@ -1,8 +1,26 @@
 import { Collection, ImportDeclaration, JSCodeshift } from 'jscodeshift'
-import { ImportDeclarationCollectionExtensions } from '../types/extended-import-declaration-collection'
-import { ExtendedImportSpecifierCollection } from '../types/extended-import-specifier-collection'
+import { ExtendedCollection } from '../types/extended-collection'
 import { ExtendedJSCodeshift } from '../types/extended-jscodeshift'
 import { once } from './once'
+import { ExtendedImportSpecifierCollection } from './register-import-specifier-collection-extensions'
+
+/**
+ * Represents an `ImportDeclaration` node collection that has access to type-specific extension methods
+ * via `utils/register-extensions.ts`
+ */
+export interface ExtendedImportDeclarationCollection
+  extends ImportDeclarationCollectionExtensions,
+    ExtendedCollection<ImportDeclaration> {}
+
+/**
+ * Extension methods that are unique to `ImportDeclaration` collections
+ */
+export interface ImportDeclarationCollectionExtensions {
+  /**
+   * Returns a collection of `ImportSpecifier` nodes with the provided name (case-sensitive)
+   */
+  findImportSpecifiersByName: (name: string) => ExtendedImportSpecifierCollection
+}
 
 const _registerImportDeclarationCollectionExtensions = (
   jscodeshift: ExtendedJSCodeshift | JSCodeshift
