+++ codemods/src/utils/register-jsx-element-collection-extensions.ts
@@ -1,7 +1,6 @@
-import { JSCodeshift, JSXAttribute, JSXSpreadAttribute } from 'jscodeshift'
+import { JSCodeshift, JSXAttribute, JSXElement, JSXSpreadAttribute } from 'jscodeshift'
 import { ExtendedCollection } from '../types/extended-collection'
 import { ExtendedJSCodeshift } from '../types/extended-jscodeshift'
-import { JSXElementCollectionExtensions, ExtendedJSXElementCollection } from '../types/extended-jsx-element-collection'
 import { flatMap } from './flat-map'
 import {
   getJsxAttributes,
@@ -11,6 +10,49 @@ import {
   isEqual
 } from './jsx-attribute-utils'
 import { once } from './once'
+
+/**
+ * Represents a `JSXElement` node collection that has access to type-specific extension methods
+ * via `utils/register-extensions.ts`
+ */
+export interface ExtendedJSXElementCollection
+  extends JSXElementCollectionExtensions,
+    Omit<ExtendedCollection<JSXElement>, 'renameTo'> {}
+
+/**
+ * Extension methods that are unique to `ImportDeclaration` collections
+ */
+export interface JSXElementCollectionExtensions {
+  /**
+   * Adds the given prop to the `JSXOpeningElement`
+   */
+  addProp: (prop: JSXAttribute | JSXSpreadAttribute) => ExtendedJSXElementCollection
+
+  /**
+   * Returns a collection of the `JSXAttribute` nodes with the provided name for the current `JSXElement` collection
+   */
+  findPropWithName: (name: string) => ExtendedCollection<JSXAttribute>
+
+  /**
+   * Returns a collection of `JSXElement` nodes that have of the provided prop name
+   */
+  findWithPropName: (name: string) => ExtendedJSXElementCollection
+
+  /**
+   * Returns a collection of `JSXElement` nodes that have spread props
+   */
+  findWithSpreadProps: () => ExtendedJSXElementCollection
+
+  /**
+   * Renames the `JSXOpeningElement` and `JSXClosingElement` (if present) to the provided value
+   */
+  renameTo: (name: string) => ExtendedJSXElementCollection
+
+  /**
+   * Removes the prop from the `JSXOpeningElement` (if present)
+   */
+  removeProp: (prop: JSXAttribute | JSXSpreadAttribute) => ExtendedJSXElementCollection
+}
 
 const _registerJSXElementCollectionExtensions = (jscodeshift: ExtendedJSCodeshift | JSCodeshift) => {
   const j = jscodeshift as ExtendedJSCodeshift
